backend/pyproject.toml
[tool.poetry]
name = "storageroomapi"
version = "0.1.0"
description = ""
authors = ["DraksPlay <draksplay@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "4.2"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.9"
djangorestframework = "^3.15.1"
drf-yasg = "^1.21.7"
django-cors-headers = "^4.3.1"
pyjwt = "^2.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
